+++ NEWS.md
@@ -8,6 +8,7 @@
   - Added finetuning for depth and layers
   - Enabled TimeGPT as a global model
   - Added support for padding time series that don't meet minimum data requirements
+  - Integrated TimeGPT into the finn agent workflow
 
 - Updated Train Model function
   - Added debug arg to trace error while training over various models